clean up zip.cpp, enable and fix compiler warnings, use equality instead of hashing for style components, compile all cpp files in sample directory, track remote branches for submodules

+++ source/workbook/workbook.cpp
@@ -637,7 +637,7 @@ std::size_t workbook::get_index(worksheet ws)
 		throw invalid_parameter();
     }
 
-    return std::distance(begin(), match);
+    return static_cast<std::size_t>(std::distance(begin(), match));
 }
 
 void workbook::create_named_range(const std::string &name, worksheet range_owner, const std::string &reference_string)
